add binding logic between author and task

+++ taskmgr/storage/task_store_db.go
@@ -16,16 +16,15 @@ type TaskStore struct {
 	Conn *pgx.Conn
 }
 
-func (db *TaskStore) GetOne(id string) (*task.Task, error) {
+func (db *TaskStore) GetOne(id, authorID string) (*task.Task, error) {
 	query := fmt.Sprintf("SELECT * FROM task WHERE id = '%s';", id)
 
 	t := task.Task{}
-	var author string
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
 
-	row := db.Conn.QueryRow(ctx, query).Scan(&t.ID, &t.Title, &t.Desc, &author)
+	row := db.Conn.QueryRow(ctx, query).Scan(&t.ID, &t.Title, &t.Desc, &t.AuthorID)
 	if row == pgx.ErrNoRows {
 		return nil, fmt.Errorf("task not found")
 	}
@@ -33,9 +32,9 @@ func (db *TaskStore) GetOne(id string) (*task.Task, error) {
 	return &t, nil
 }
 
-func (db *TaskStore) GetTasks() []*task.Task {
+func (db *TaskStore) GetTasks(authorID string) []*task.Task {
 	var fetchedTasks []*task.Task
-	query := "SELECT * FROM task"
+	query := fmt.Sprintf("SELECT * FROM task WHERE author_id = '%s'", authorID)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -49,9 +48,8 @@ func (db *TaskStore) GetTasks() []*task.Task {
 
 	for rows.Next() {
 		r := &task.Task{}
-		var author string
 
-		err := rows.Scan(&r.ID, &r.Title, &r.Desc, &author)
+		err := rows.Scan(&r.ID, &r.Title, &r.Desc, &r.AuthorID)
 		if err != nil {
 			log.Printf("%v", err)
 		}
@@ -62,48 +60,34 @@ func (db *TaskStore) GetTasks() []*task.Task {
 }
 
 func (db *TaskStore) New(task *task.Task) uuid.UUID {
-	// To be removed when proper user logic is implemented
-	query := "SELECT id FROM author WHERE email = 'marco@rossi.com';"
-	var authorID string
+	query := fmt.Sprintf("INSERT INTO task (author_id, title, description) VALUES ('%s', '%s', '%s');", task.AuthorID, task.Title, task.Desc)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
 
-	err := db.Conn.QueryRow(ctx, query).Scan(&authorID)
-	if err == pgx.ErrNoRows {
-		log.Printf("Couldn't find a match: %v", err)
-	}
-
-	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	query = fmt.Sprintf("INSERT INTO task (title, description, author_id) VALUES ('%s', '%s', '%s');", task.Title, task.Desc, authorID)
-	_, err = db.Conn.Exec(ctx, query)
+	_, err := db.Conn.Exec(ctx, query)
 	if err != nil {
 		log.Printf("Could not insert new record into the database %v", err)
 		return uuid.Nil
 	}
 
-	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-
-	query = fmt.Sprintf("SELECT id FROM task WHERE title = '%s'", task.Title)
-  
-	err = db.Conn.QueryRow(ctx, query).Scan(&task.ID)
-	if err == pgx.ErrNoRows {
-		log.Printf("Couldn't find a match: %v", err)
+	query = fmt.Sprintf("SELECT id FROM task WHERE title='%s'", task.Title)
+	row := db.Conn.QueryRow(ctx, query).Scan(&task.ID)
+	if row == pgx.ErrNoRows {
+		log.Printf("No task was found with ID: %v", err)
+		return uuid.Nil
 	}
 
 	return task.ID
 }
 
-func (db *TaskStore) Update(id, title, desc string) (*task.Task, error) {
+func (db *TaskStore) Update(id, title, desc, authorID string) (*task.Task, error) {
 	UUID, err := uuid.Parse(id)
 	if err != nil {
 		return nil, fmt.Errorf("invalid uuid: %s", id)
 	}
 
-	query := fmt.Sprintf(`UPDATE task SET title = '%s', description = '%s' WHERE id = '%s';`, title, desc, id)
+	query := fmt.Sprintf(`UPDATE task SET title = '%s', description = '%s' WHERE id = '%s' and author_id = '%s';`, title, desc, id, authorID)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -132,13 +116,13 @@ func (db *TaskStore) Update(id, title, desc string) (*task.Task, error) {
 	return &updatedTask, nil
 }
 
-func (db *TaskStore) Delete(id string) error {
+func (db *TaskStore) Delete(id, authorID string) error {
 	UUID, err := uuid.Parse(id)
 	if err != nil {
 		return fmt.Errorf("invalid uuid: %s", id)
 	}
 
-	query := fmt.Sprintf("DELETE FROM task WHERE id = '%s';", id)
+	query := fmt.Sprintf("DELETE FROM task WHERE id = '%s' and author_id = '%s';", id, authorID)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
@@ -156,9 +140,9 @@ func (db *TaskStore) Delete(id string) error {
 	return nil
 }
 
-func (db *TaskStore) BulkImport(tasks []*task.Task) {
+func (db *TaskStore) BulkImport(tasks []*task.Task, authorID string) {
 	for _, t := range tasks {
-		query := fmt.Sprintf("INSERT INTO task (title, description) VALUES ('%s', '%s');", t.Title, t.Desc)
+		query := fmt.Sprintf("INSERT INTO task (title, description, author_id) VALUES ('%s', '%s', '%s');", t.Title, t.Desc, authorID)
 
 		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 		defer cancel()
